Refactor extracted script linter from TitleTemplateLinter to TitleScriptLinter
- Removed the validateScript method and its associated logic from TitleTemplateLinter.
- Updated manual tests to include TitleScriptLinter for script validation.
- Cleaned up the TitleTemplateLinter class to focus solely on template validation.
- Adjusted test cases to reflect the removal of script validation from TitleTemplateLinter.
